tests: extract the Corkami corpus helper into its own module
Importing test_corkami.py runs _init_tests() at module scope, which
downloads the corpus and registers a test method per file. Move the
corpus location, download, and extraction into tests/corkami_corpus.py
so other test modules can read a sample without that side effect, and
add a read() helper that returns one file's bytes.
